chore(core): rename useRenderingContext hook file
