+++ packages/sanity/src/core/canvas/actions/LinkToCanvas/LinkToCanvasAction.tsx
@@ -11,7 +11,7 @@ import {useSchema} from '../../../hooks/useSchema'
 import {useTranslation} from '../../../i18n/hooks/useTranslation'
 import {usePerspective} from '../../../perspective/usePerspective'
 import {isReleaseDocument} from '../../../releases/store/types'
-import {useRenderingContext} from '../../../store/renderingContext/useIsInRenderContext'
+import {useRenderingContext} from '../../../store/renderingContext/useRenderingContext'
 import {getDraftId, getPublishedId} from '../../../util/draftUtils'
 import {canvasLocaleNamespace} from '../../i18n'
 import {useCanvasCompanionDoc} from '../useCanvasCompanionDoc'
